+++ src/components/Admin/UserPage.tsx
@@ -1,9 +1,9 @@
 import React, { FC } from 'react';
-import { Button, Col, Container, Row } from 'react-bootstrap';
+import { Button, Col, Row } from 'react-bootstrap';
 import { Link, Route, Switch, useRouteMatch } from 'react-router-dom';
 import { useUsersQuery } from '../../generated/graphql';
 import { UserModel } from '../../models/User';
-import { UserInput, EditMode } from './UserInput';
+import { EditMode, UserInput } from './UserInput';
 import UserList from './UserList';
 
 export const UserPage: FC = () => {
@@ -35,22 +35,24 @@ export const UserPage: FC = () => {
       {loading ? (
         <h3>Loading...</h3>
       ) : (
-        <Container>
-          <Switch>
-            <Route exact path={`${path}`}>
-              <UserList users={users} />
-            </Route>
-            <Route exact path={`${path}/new`}>
-              <UserInput key="new" users={users} editMode={EditMode.new} onSave={handleSaveUser} />
-            </Route>
-            <Route exact path={`${path}/:userId/edit`}>
-              <UserInput key="edit" users={users} editMode={EditMode.edit} onSave={handleSaveUser} />
-            </Route>
-            <Route exact path={`${path}/:userId`}>
-              <UserInput key="edit" users={users} onSave={handleSaveUser} />
-            </Route>{' '}
-          </Switch>
-        </Container>
+        <Row>
+          <Col>
+            <Switch>
+              <Route exact path={`${path}`}>
+                <UserList users={users} />
+              </Route>
+              <Route exact path={`${path}/new`}>
+                <UserInput key="new" users={users} editMode={EditMode.new} onSave={handleSaveUser} />
+              </Route>
+              <Route exact path={`${path}/:userId/edit`}>
+                <UserInput key="edit" users={users} editMode={EditMode.edit} onSave={handleSaveUser} />
+              </Route>
+              <Route exact path={`${path}/:userId`}>
+                <UserInput key="edit" users={users} onSave={handleSaveUser} />
+              </Route>{' '}
+            </Switch>
+          </Col>
+        </Row>
       )}
     </>
   );
